refactor(mfa): colocate event executors with view config in the test
The event executors and the state assertions together form the UI-walk
test configuration, so move mfaEventExecutors out of the render harness
and next to testConfig in the test file. The harness now only mounts the
UI and exposes getMfaControls for the INIT executor.
